+++ examples/better-auth/package.json
@@ -16,6 +16,7 @@
     "react": "19.2.0",
     "react-dom": "19.2.0",
     "sign-in-with-stacks": "workspace:*",
+    "siws-better-auth": "workspace:*",
     "zod": "4.1.12"
   },
   "devDependencies": {
