+++ examples/better-auth/src/lib/auth-client.ts
@@ -1,5 +1,5 @@
 import { createAuthClient } from "better-auth/react";
-import { siwsClient } from "./siws";
+import { siwsClient } from "siws-better-auth";
 
 export const authClient = createAuthClient({
   plugins: [siwsClient()],
