+++ examples/better-auth/src/lib/auth.ts
@@ -1,5 +1,5 @@
 import { betterAuth } from "better-auth";
-import { siws } from "./siws";
+import { siws } from "siws-better-auth";
 
 export const auth = betterAuth({
   plugins: [
