--- examples/better-auth/src/lib/siws/index.ts
@@ -1,2 +1,0 @@
-export * from "./client";
-export * from "./plugin";
