+++ packages/siws-better-auth/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./client.js";
+export * from "./plugin.js";
