+++ packages/siws-better-auth/src/schema.ts
