+++ packages/siws-better-auth/src/types.ts
