+++ packages/siws-better-auth/tsdown.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig, type UserConfig, type UserConfigFn } from "tsdown";
+
+const config: UserConfigFn | UserConfig = defineConfig({
+  entry: ["src/index.ts"],
+  sourcemap: true,
+  clean: true,
+  dts: true,
+  publint: true,
+  platform: "neutral",
+});
+
+export default config;
